\name{ex1}
\docType{data}
\alias{ex1}
\title{
Example 1 Data Set
}
\description{
\code{ex1} contains 100 simulated concentrations during the times between 0 and 4, using the parameters Beta = 5, Q = 13.8, G = 351.5, VN = pi*10^-3, VF = 3.8, Tau_N = 1, Tau_NF = 0.5 and Tau_F = 0.64. 
}
\seealso{
\code{\link{ex2}}
}
\author{
Joao Vitor Dias Monteiro, Sudipto Banerjee and Gurumurthy Ramachandran.

Maintainer: Joao Vitor Dias Monteiro <monte092@umn.edu>
}
